Finally got the async waveform generation going and it's awesome!
