+++ src/ui/molecules/Track/AudioRegion/AudioRegion.tsx
@@ -37,6 +37,7 @@ function AudioRegion() {
   useRegionDawRecordingSync();
   useRegionScheduler();
 
+
   const isDuplicating = isPressed('alt') && isMoving;
 
   return (
